+++ src/comps/Toggle.svelte
@@ -13,7 +13,7 @@
   export { className as class };
   export let onVisibility = () => {};
 
-  let isVisible = true;
+  let isVisible = false;
 
   const onClick = () => {
     if (!alwaysopen) {
@@ -25,8 +25,9 @@
 
   onMount(() => {
     const lsValue = localStore.visibility(title);
-    if (!lsValue || open || alwaysopen) {
+    if (lsValue || open || alwaysopen) {
       isVisible = lsValue || open || alwaysopen;
+      onVisibility(isVisible); // V4IY
     }
   });
 </script>