+++ src/stores/global-data.ts
@@ -1,6 +1,9 @@
 import { writable } from 'svelte/store';
-
 import dataJSON from '../../data/data.json';
 
-export const storedGlobalData = writable({});
-storedGlobalData.set(dataJSON);
+const { subscribe, set, update } = writable(dataJSON);
+export const storedGlobalData = {
+  subscribe,
+  set,
+  update,
+};